packages/indexed-core/pyproject.toml
[project]
name = "indexed-core"
version = "0.1.0"
description = "Core library for document indexing and semantic search with FAISS and sentence-transformers"
readme = "README.md"
authors = [
    { name = "Lennard Zündorf", email = "l.p.zuendorf@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.26",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "unstructured[all-docs]>=0.18.5",
    "pydantic-settings>=2.2.1",
    "platformdirs>=4.2.0",
    "tomlkit>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
